+++ lib/features/settings/custom_pattern_screen.dart
@@ -356,4 +356,3 @@ class _ZoneReorderTile extends StatelessWidget {
     );
   }
 }
-